requirements/dev.txt
-r base.txt
# Testing 
# -----------------------
mypy == 1.3.0
pytest == 7.3.1

# Code Quality
# -----------------------
flake8 == 6.0.0
coverage == 7.2.7

# Djanngo extensions 
# ---------------------------------
django-debug-toolbar == 4.1.0